fix: removed new and rework keywords
