+++ client/src/components/game/phases/eventsPhase/NewCardEvent.vue
@@ -44,5 +44,5 @@ export default class CardEvent extends Vue {
 </script>
 
 <style lang="scss" scoped>
-@import '@port-of-mars/client/stylesheets/game/phases/eventsPhase/NewCardEvent.scss';
+@import '@port-of-mars/client/stylesheets/game/phases/eventsPhase/CardEvent.scss';
 </style>
