+++ client/src/components/global/UserHeader.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="user-header">
     <div class="problem">
-      <a href="mailto:portmars@asu.edu?subject=[port-of-mars] New Issue Submission" target="_blank">[ <span>BETA</span> ] <span>Report a Problem</span></a>
+      <a href="mailto:portmars@asu.edu?subject=[port-of-mars]  Issue Submission" target="_blank">[ <span>BETA</span> ] <span>Report a Problem</span></a>
     </div>
     <div class="logout">
       <button
