+++ client/src/components/root/GameboardContainer.vue
@@ -1,22 +1,22 @@
 <template>
   <div class="gameboardcontainer container">
     <div class="gameboardstatusbar row">
-      <NewSystemHealth />
+      <SystemHealth />
     </div>
     <div class="gameboardcontainerrow row no-gutters">
       <div class="gameboard-wrapper col-9">
         <ProfileMenu />
         <div class="gameboardtop row">
-          <NewGameBoardTop />
+          <GameBoardTop />
         </div>
         <div class="gameboardbottom row">
-          <NewGameBoardBottom />
+          <GameBoardBottom />
         </div>
         <ActiveEventsPopup />
         <InventoryPopup />
       </div>
       <div class="chatwrapper col-3">
-        <NewChat/>
+        <Chat/>
       </div>
     </div>
 </div>
@@ -24,23 +24,23 @@
 
 <script lang="ts">
 import { Vue, Component } from 'vue-property-decorator';
-import NewGameBoardTop from '@port-of-mars/client/components/game/static/NewGameBoardTop.vue';
-import NewGameBoardBottom from '@port-of-mars/client/components/game/phases/NewGameBoardBottom.vue';
+import GameBoardTop from '@port-of-mars/client/components/game/static/GameBoardTop.vue';
+import GameBoardBottom from '@port-of-mars/client/components/game/phases/GameBoardBottom.vue';
 import ProfileMenu from '@port-of-mars/client/components/game/static/popups/ProfileMenu.vue';
 import ActiveEventsPopup from '@port-of-mars/client/components/game/static/popups/ActiveEventsPopup.vue';
 import InventoryPopup from '@port-of-mars/client/components/game/static/popups/InventoryPopup.vue';
-import NewChat from '@port-of-mars/client/components/game/static/chat/NewChat.vue'
-import NewSystemHealth from '@port-of-mars/client/components/game/static/systemHealth/NewSystemHealth.vue';
+import Chat from '@port-of-mars/client/components/game/static/chat/Chat.vue'
+import SystemHealth from '@port-of-mars/client/components/game/static/systemHealth/SystemHealth.vue';
 
 @Component({
   components: {
-    NewGameBoardTop,
-    NewGameBoardBottom,
+    GameBoardTop,
+    GameBoardBottom,
     ProfileMenu,
     ActiveEventsPopup,
     InventoryPopup,
-    NewChat,
-    NewSystemHealth,
+    Chat,
+    SystemHealth,
   }
 })
 export default class GameboardContainer extends Vue {}
