+++ client/src/stylesheets/game/phases/GameBoardBottom.scss
