+++ client/src/stylesheets/game/phases/discardPhase/ContainerDiscard.scss
