+++ client/src/stylesheets/game/phases/eventsPhase/CardEvent.scss
