+++ client/src/stylesheets/game/phases/eventsPhase/ContainerEvents.scss
@@ -1,75 +1,68 @@
-.container-events {
-  @include expand;
-  padding: 0.5rem;
-
-  .events-row {
-    @include expand;
-    margin: 0;
-    @include row;
-  }
-
-  .info {
+.events-container{
+    display: flex;
+    justify-content: space-between;
+    min-height: 0;
     height: 100%;
-    @include column-3;
-    padding: 0;
-    margin-right: 0.5rem;
-    @include make-column-and-center;
 
-    .topbar {
-      height: 10%;
-      width: 100%;
-      margin: 0 0 0.5rem 0;
-      padding: 0.25rem 1rem;
-      @include make-center;
-      background-color: $new-space-orange;
+    .section-text{
+        height: 3rem;
+        width: 100%;
+        @include make-column-and-center;
+        
+        
+        background-color: $new-space-orange;
 
-      p {
-        margin: 0;
-        color: $space-gray;
-        font-size: $font-med;
-      }
+        p{
+            margin-top:0.65rem;
+            font-weight: bold;
+            color: black;
+            
+        }
     }
 
-    .content {
-      flex-grow: 1;
-      width: 100%;
-      padding: 0.5rem;
-      margin: 0;
-      @include make-column-and-center;
-      background-color: $space-white-opaque-1;
+    .event-deck-container{
+        @include column-3;
+        background-color: $space-white-opaque-1;
+        overflow-y: scroll;
+        @include disable-scroll;
+        height: 100%;
 
-      p {
-        margin: 0.5rem 0;
-        color: $new-space-orange;
-        font-size: $font-med;
-        text-align: center;
-      }
+        .events{
+            height: 100%;
+            
 
-      .name-wrapper,
-      .effect-wrapper {
-        margin: 0.5rem;
-        font-size: $font-med;
-      }
-
-      .name,
-      .effect {
-        color: $space-white;
-        font-size: $font-small;
-      }
+            .event-container{
+                width: 95%;
+                margin: 1rem auto;
+            }
+        }
     }
-  }
 
-  .actions {
-    height: 100%;
-    flex-grow: 1;
-    padding: 0.5rem;
-    background-color: $space-white-opaque-1;
-    color: $space-white;
+    .active-event-container{
+        flex: 1 1 0;
+        height: 100%;
+        margin: 0 0.5rem;
+        background-color: $space-white-opaque-1;
+        color:white;
 
-    .outer-wrapper {
-      @include expand;
-      overflow-y: scroll;
-      @include disable-scroll;
+        .event-holder{
+
+            .event-info{
+                
+                .event-title{
+                    color: $new-space-orange;
+                    font-size: $font-large;
+                    font-style: italic;
+                    margin-left:0.5rem;
+                    p{
+                        width:35rem;
+                        border-bottom: solid 0.025rem white;
+                    }
+                }
+            }
+
+        }
+
     }
-  }
+
 }
