+++ client/src/stylesheets/game/phases/purchasePhase/ContainerPurchase.scss
