+++ client/src/stylesheets/game/phases/tradePhase/TradeOptions.scss
