+++ client/src/stylesheets/game/phases/tradePhase/TradeRequest.scss
