+++ client/src/stylesheets/game/static/GameBoardTop.scss
